feat: custom Cypress command for creating variants and testsets
