+++ agenta-web/cypress/support/commands/index.ts
@@ -6,6 +6,7 @@ declare global {
     namespace Cypress {
         interface Chainable {
             clickLinkAndWait(selector: string): Chainable<void>
+            createVariantsAndTestsets(): Chainable<void>
         }
     }
 }
